# refactor connection ID manager to use a slice instead of a linked list