The list of connection IDs will always be short (less than 10 items), so there's no real reason to use a linked list. The benchmark proves that using a slice is indeed faster.

This will help with #4337.

```
❯ benchstat old.txt new.txt
name                       old time/op    new time/op    delta
ConnIDManagerReordered-16     131ns ± 8%     114ns ± 2%  -13.13%  (p=0.008 n=5+5)
ConnIDManagerInOrder-16       122ns ± 1%     113ns ± 3%   -7.21%  (p=0.008 n=5+5)

name                       old alloc/op   new alloc/op   delta
ConnIDManagerReordered-16      136B ± 0%      152B ± 0%  +11.76%  (p=0.008 n=5+5)
ConnIDManagerInOrder-16        136B ± 0%      152B ± 0%  +11.76%  (p=0.008 n=5+5)

name                       old allocs/op  new allocs/op  delta
ConnIDManagerReordered-16      3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.008 n=5+5)
ConnIDManagerInOrder-16        3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.008 n=5+5)
```